Fix apps-init (#55)
apps-init will fail due to the fact it uses a "lets check for all
dependencies" type behaviour. This commit gets rid of this type of dependency
checking. If an external tool is to be used the the command then the command
that uses it should report directly to the user.

Signed-off-by: Jorge Alarcon Ochoa <alarcj137@gmail.com>