Refactored Dockerfiles and make targets
To support rapid development, we need to be able to turn off the Docker cache
for the installation of AgavePy and other late-stage assets. But, we rarely
need to repeat all the early stage dependency and config. So, the build
process was refactored so that Py2/3 base images (cli-base:py3 and cli-base:py2)
are build via 'make bases'. These local images are then used as source of the
public and testing images.